Move components folder into src/
