+++ src/pages/kontakt.js
@@ -1,4 +1,4 @@
-import Page from "../../components/Page";
+import Page from "../components/Page";
 
 const Contact = () => {
   return <Page title="Kontaktinfo"></Page>;
