+++ src/pages/om.js
@@ -1,4 +1,4 @@
-import Page from "../../components/Page";
+import Page from "../components/Page";
 
 const About = () => {
   return <Page title="Om oss"></Page>;
